Styleguide compliance and readme consolidation.
