+++ .env.example
@@ -2,7 +2,7 @@
 # These are environment-specific settings that change between dev/staging/prod
 
 # Option 1: Service Account Key File (Recommended)
-GOOGLE_APPLICATION_CREDENTIALS=/path/to/service-account-key.json
+GOOGLE_APPLICATION_CREDENTIALS=./.serviceAccount.json
 
 # Option 2: Individual Credentials
 # GCP_CLIENT_EMAIL=your-service-account@your-project.iam.gserviceaccount.com
