+++ .secrets.example
@@ -3,19 +3,23 @@
 # DO NOT commit .secrets to version control!
 
 # This file is used by act to simulate GitHub repository secrets when testing
-# workflows locally. It combines all secrets needed by the workflows.
+# workflows locally.
 
 # Required for npm publishing (get from npmjs.com)
 NPM_TOKEN=your-npm-automation-token-here
 
-# BigQuery Configuration for Integration Tests
-# These would normally be set as GitHub repository secrets
-# For local act testing, we provide them here
-BIGQUERY_CREDENTIALS={"type":"service_account","project_id":"your-project",...}
-BIGQUERY_PROJECT_ID=your-gcp-project-id
-BIGQUERY_DATASET=kysely_bigquery_test
+# For act testing, you may need to provide BigQuery secrets that GitHub Actions expects
+# These should match what's configured in your GitHub repository secrets
+#
+# Option 1: Read from your .serviceAccount.json file and paste here:
+# BIGQUERY_CREDENTIALS=<paste entire service account JSON>
+#
+# Option 2: Use environment variables from .env (recommended):
+# The test script can be modified to pass .env values as secrets to act
+#
+# Example:
+# BIGQUERY_CREDENTIALS={"type":"service_account",...}
+# BIGQUERY_PROJECT_ID=your-project-id
+# BIGQUERY_DATASET=your_test_dataset
 
 # Note: GITHUB_TOKEN is automatically provided by act, no need to set it here
-
-# For actual local development (running tests directly), use .env file
-# See .env.example for local development configuration
