+++ src/index.ts
@@ -1,7 +1,7 @@
-export * from './BigQueryAdapter';
-export * from './BigQueryConnection';
-export * from './BigQueryDialect';
-export * from './BigQueryDriver';
-export * from './BigQueryIntrospector';
-export * from './BigQueryCompiler';
-export * from './json-column-detector';
+export * from './bigQueryAdapter';
+export * from './bigQueryConnection';
+export * from './bigQueryDialect';
+export * from './bigQueryDriver';
+export * from './bigQueryIntrospector';
+export * from './bigQueryCompiler';
+export * from './jsonColumnDetector';
