+++ src/jsonColumnDetector.ts
