+++ tests/bigqueryJsonConfig.integration.test.ts
