+++ tests/bigqueryJsonStringColumns.integration.test.ts
