+++ tsconfig.json
@@ -14,7 +14,11 @@
     "outDir": "dist",
     "alwaysStrict": true,
     "preserveConstEnums": true,
-    "declaration": true
+    "declaration": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noImplicitReturns": true,
+    "noFallthroughCasesInSwitch": true
   },
   "include": ["src"],
   "exclude": ["node_modules", "dist", "test"]
